remove legacy javascript modal logic
