+++ server/app/assets/javascripts/admin_entry_point.ts
@@ -20,7 +20,6 @@ import * as adminValidation from './admin_validation'
 import * as apiDocs from './api_docs'
 import * as devIcons from './dev_icons'
 import * as map from './mapquestion/map'
-import * as modal from './modal'
 import * as northStarModal from './north_star_modal'
 import * as questionBank from './questionBank'
 import PreviewController, * as preview from './preview'
@@ -72,7 +71,6 @@ function initializeEverything(): void {
   adminValidation.init()
   apiDocs.init()
   devIcons.init()
-  modal.init()
   northStarModal.init()
   questionBank.init()
   preview.init()
