+++ server/app/assets/javascripts/applicant_entry_point.ts
@@ -9,7 +9,6 @@ import * as enumerator from './enumerator'
 import * as radio from './radio'
 import * as toast from './toast'
 import * as map from './mapquestion/map'
-import * as modal from './modal'
 import * as northStarModal from './north_star_modal'
 import * as fileUpload from './file_upload'
 import * as azureDelete from './azure_delete'
@@ -37,7 +36,6 @@ window.addEventListener('load', () => {
   radio.init()
   toast.init()
   map.init()
-  modal.init()
   northStarModal.init()
   fileUpload.init()
   azureDelete.init()
